+++ scripts/quick-journal-page-clipboard.js
@@ -1,7 +1,36 @@
-//CONFIG.debug.hooks = true
-//console.log("CLI: Module Quick-Journal-Page-Clippboard has started");
-Hooks.once("ready", () => {
-  game.settings.register("quick-journal-page-clipboard", "removeGMSecrets", {
+CONFIG.debug.hooks = true
+//console.log("QJPC: Module Quick-Journal-Page-Clippboard has started");
+
+import { convertHtmlToMarkdown } from "./render-markdown.js";
+
+let sheetObject = null;
+
+
+
+Hooks.once("init", () => {
+
+  
+//TODO: restructure trannslations to nested JSON AND to automatic settings translation for name and hint
+// new settings setup is possible:
+// {Scope}.settings.{SettingSubkey}.Name
+// {Scope}.settings.{SettingSubkey}.Hint
+
+/* JSON structure for this 
+{
+  "quick-journal-page-clipboard": {
+    "settings": {
+      "returnMDtext": {
+        "Name": "Return Markdown Text",
+        "Hint": "If enabled, text copied from the journal will maintain its Markdown formatting."
+      }
+    }
+  }
+}
+*/
+
+// now name and hint can be removed as propperties
+
+game.settings.register("quick-journal-page-clipboard", "removeGMSecrets", {
     name: game.i18n.localize("QJPC.settings.removeGMSecrets.name"),
     hint: game.i18n.localize("QJPC.settings.removeGMSecrets.hint"),
     scope: "world",
@@ -14,7 +43,7 @@ Hooks.once("ready", () => {
   game.settings.register("quick-journal-page-clipboard", "isCuratedText", {
     name: game.i18n.localize("QJPC.settings.isCuratedText.name"),
     hint: game.i18n.localize("QJPC.settings.isCuratedText.hint"),
-    scope: "world",
+    scope: "user",
     config: true,
     type: Boolean,
     default: true,
@@ -24,7 +53,7 @@ Hooks.once("ready", () => {
   game.settings.register("quick-journal-page-clipboard", "isEmptyLineText", {
     name: game.i18n.localize("QJPC.settings.isEmptyLineText.name"),
     hint: game.i18n.localize("QJPC.settings.isEmptyLineText.hint"),
-    scope: "world",
+    scope: "user",
     config: true,
     type: Boolean,
     default: true,
@@ -34,7 +63,18 @@ Hooks.once("ready", () => {
   game.settings.register("quick-journal-page-clipboard", "isPageTitle", {
     name: game.i18n.localize("QJPC.settings.isPageTitle.name"),
     hint: game.i18n.localize("QJPC.settings.isPageTitle.hint"),
-    scope: "world",
+    scope: "user",
+    config: true,
+    type: Boolean,
+    default: false,
+    restricted: false
+  });
+
+  /*
+    game.settings.register("quick-journal-page-clipboard", "returnMDtext", {
+    name: game.i18n.localize("QJPC.settings.returnMDtext.name"),
+    hint: game.i18n.localize("QJPC.settings.returnMDtext.hint"),
+    scope: "user",
     config: true,
     type: Boolean,
     default: false,
@@ -44,38 +84,127 @@ Hooks.once("ready", () => {
   game.settings.register("quick-journal-page-clipboard", "returnHTMLtext", {
     name: game.i18n.localize("QJPC.settings.returnHTMLtext.name"),
     hint: game.i18n.localize("QJPC.settings.returnHTMLtext.hint"),
-    scope: "world",
+    scope: "user",
     config: true,
     type: Boolean,
     default: false,
     restricted: false
-  });
+  });*/
+
+game.settings.register("quick-journal-page-clipboard", "outputFormat", {
+  name: game.i18n.localize("QJPC.settings.outputFormat.name"),
+  hint: game.i18n.localize("QJPC.settings.outputFormat.hint"),
+  scope: "user",
+  config: true,
+  type: String,
+  choices: {
+    plaintext: game.i18n.localize("QJPC.settings.plaintext.name"),
+    markdown: game.i18n.localize("QJPC.settings.returnMDtext.name"),
+    html: game.i18n.localize("QJPC.settings.returnHTMLtext.name")
+  },
+  default: "plaintext",
+  restricted: false
 });
+
+
+
+  game.keybindings.register("quick-journal-page-clipboard", "showNotification", {
+  name: "My Settings Keybinding",
+  hint: "A description of what will occur when the Keybinding is executed.",
+  editable: [
+    {
+      key: "KeyC",
+      modifiers: ["Control","Alt"]
+    }
+  ],
+onDown: async () => { 
+
+  // Check if a journal sheet is currently open and on TOP (not focused) - this does not allways play any role as sheetObject can be filled allready
+  // but it will irritate the user if he can fill the clipboard anyway if no journal is open
+    if(!sheetObject) {
+    ui.notifications.warn(game.i18n.localize("QJPC.notifications.noJournalOpen"));
+    return;
+  }
+
+
+const isAJournalCurrentlyActive = getAbsoluteTopJournal();
+
+//console.log("QJPC: top window: ",isAJournalCurrentlyActive)
+// Only go on if there is a journal currently opened and on top
+if (!isAJournalCurrentlyActive) {
+  ui.notifications.warn(game.i18n.localize("QJPC.notifications.noJournalOpen"));
+  return;
+}
+
+  try {
+    await getClipboardText(sheetObject);
+
+    //Get the export text
+    //const text = await exportJournalText(sheetObject);
+    //await navigator.clipboard.writeText(text);
+    //if(text!==""){ui.notifications.info(game.i18n.localize("QJPC.notifications.copied"));}       
+  } catch (error) {
+    console.error(game.i18n.localize("QJPC.warning.error"), error);
+    ui.notifications.error(game.i18n.localize("QJPC.warning.error"));
+  }
+},
+  onUp: () => {},
+  restricted: false,             // Restrict this Keybinding to gamemaster only?
+  //reservedModifiers: ["Alt"],  // On ALT, the notification is permanent instead of temporary
+  precedence: CONST.KEYBINDING_PRECEDENCE.NORMAL
+});
+
+
+/*
+Hooks.on("renderSettingsConfig", (changed) => {
+  console.log("QJPC: settingsUpdate: ",changed)
+  if (
+    changed.has("quick-journal-page-clipboard.returnMDtext") ||
+    changed.has("quick-journal-page-clipboard.returnHTMLtext")
+  ) {
+    ensureMutualExclusion();
+  }
+});*/
+
+});
+
+
 
 Hooks.on("getHeaderControlsJournalEntrySheet", (sheet, buttons) => {
   const isJournal = sheet?.document?.documentName ?? "noJournal";
 
-  if(!sheet.document || !isJournal==="JournalEntry") {return;};
+  if(!sheet.document || isJournal!=="JournalEntry") {return;};
   //console.log("qjpc: Sheet: ",sheet)
-    
+    sheetObject = sheet; // store the actual sheet into sheetObject to be able to access it with key combination
     const actionName = "textToClipboard";
     buttons.push({
         icon: "fa-solid fa-clipboard",
         label: game.i18n.localize("QJPC.controls.clipboard.label"),
         action: actionName,
         class: "qjpc",
-        visible: true
+        visible: true,
+  
     });
+
 
     //if the action does not yet exist create it
     if (!sheet.options.actions[actionName]) {
-        sheet.options.actions[actionName] = async function(event, target) {
+      
+            sheet.options.actions[actionName] = async (event, target) => {
+             await getClipboardText(sheet);
+    }
+        /*
+        async function(event, target) {
          
       const text = await exportJournalText(sheet);
+
+      if(isMarkdown){
+        text = convertHtmlToMarkdown(text);
+      }
        
       await navigator.clipboard.writeText(text); // browser standard for putting text into the clipboard 
       if(text!==""){ui.notifications.info(game.i18n.localize("QJPC.notifications.copied"));}
-        };
+        };*/
     }
 });
 
@@ -90,13 +219,17 @@ async function exportJournalText(sheet) {
   const parts = [];
   
   const isPageTitle = game.settings.get("quick-journal-page-clipboard", "isPageTitle");
-  const returnHTML = game.settings.get("quick-journal-page-clipboard", "returnHTMLtext");
+  //const returnHTML = game.settings.get("quick-journal-page-clipboard", "returnHTMLtext");
+  //const isMarkdown = game.settings.get("quick-journal-page-clipboard", "returnMDtext");
+  const radioButtonFormat = game.settings.get("quick-journal-page-clipboard", "outputFormat");
+  const isMarkdown = radioButtonFormat === "markdown";
+  const returnHTML = radioButtonFormat === "html";
 
   for (const p of exportPages) {
     let text = "";
     //console.log("qjpc: Title: ",p.name)
     if(isPageTitle){
-      if(returnHTML){
+      if(returnHTML || isMarkdown){
     text = text + "<h1>" + p.name + "</h1>" + "</br>"+"\n"+"</br>"+"\n";
     } else {
       text = text +  p.name + "\n\n";
@@ -132,8 +265,8 @@ function getPagesToExport(sheet, activePages) {
 }
 
 function isTextPage(page) {
-  const type = page.type ?? page.documentName ?? "";
-  return type === "text" || type === "markdown" || type === "html" || type === "prosemirror";
+  const type = page.type ?? "";
+  return type === "text" || type === "markdown"; // TODO: check if markdown has it's own type
 }
 
 async function extractPlainTextFromPage(page) {
@@ -152,11 +285,15 @@ async function extractPlainTextFromPage(page) {
   }
 
 let resultingText = "";
-const returnHTML = game.settings.get("quick-journal-page-clipboard", "returnHTMLtext");
+//const returnHTML = game.settings.get("quick-journal-page-clipboard", "returnHTMLtext");
+//const returnMD = game.settings.get("quick-journal-page-clipboard", "returnMDtext");
+const radioButtonFormat = game.settings.get("quick-journal-page-clipboard", "outputFormat");
+  const isMarkdown = radioButtonFormat === "markdown";
+  const returnHTML = radioButtonFormat === "html";
 const isCuratedText = game.settings.get("quick-journal-page-clipboard", "isCuratedText");
 const isEmptyLine = game.settings.get("quick-journal-page-clipboard", "isEmptyLineText");
 
-if (returnHTML) {
+if (returnHTML || isMarkdown) {
     // return the inner HTML as a string
     resultingText = doc.body.innerHTML || "";
   } else {
@@ -185,3 +322,81 @@ function normalizeText(text) {
     .replace(/\n{3,}/g, "\n\n")       // collapse 3+ newlines to 2
     .trim();
 }
+
+function getAbsoluteTopJournal() {
+  // Gather EVERY open window in Foundry (Legacy + ApplicationV2) - we need this to check for ALL open windows
+  const allWindows = [
+    ...Object.values(ui.windows), 
+    ...foundry.applications.instances.values()
+  ];
+
+  // Filter down to only windows that are actively rendered on screen and thus seem active for the user
+  const renderedWindows = allWindows.filter(app => app.rendered);
+
+  // If nothing is open at all, exit and return null/nothing
+  if (renderedWindows.length === 0) return null;
+
+  //Sort ALL open windows by their DOM z-index (highest first). We grab the dom because that is what the user really sees
+  // Also this will work for APPv1 and APPv2
+  //Check if an element is a raw HTML element? If yes, use it directly. If no, grab the first item inside the jQuery wrapper (a.element?.[0])."
+  //TODO: if that is the best way and state management of v2 is not truly only an internal thing
+  const sortedWindows = renderedWindows.sort((a, b) => {
+    const elA = a.element instanceof HTMLElement ? a.element : a.element?.[0];
+    const elB = b.element instanceof HTMLElement ? b.element : b.element?.[0];
+
+    //Parse the css style for the zIndex
+    const zA = elA ? parseInt(window.getComputedStyle(elA).zIndex) || 0 : 0;
+    const zB = elB ? parseInt(window.getComputedStyle(elB).zIndex) || 0 : 0;
+
+    return zB - zA; 
+  });
+
+  // Inspect the absolute top-most window
+  const topWindow = sortedWindows[0];
+
+  // Check if this top-most window is a JournalEntry class which was formerly JournalSheet (for APPv1 only?) 
+  // or has a constructorname that includes at least Journal (if it was changed by any other app?/for APPv2
+   
+  const isTopWindowAJournal = topWindow instanceof JournalEntry || 
+                              topWindow.constructor.name.includes("Journal");
+
+                              //foundry.appv1.sheets.JournalSheet
+
+  // Only return the journal if it wins the z-index battle against ALL other apps
+  // if isTopWindowAJournal is true we either return the APPv1 or APPv2 object as topWindow. If it is false we return null
+  return isTopWindowAJournal ? topWindow : null;
+}
+
+  async function getClipboardText(sheet) {
+      //const isMarkdown = game.settings.get("quick-journal-page-clipboard", "returnMDtext");
+      const radioButtonFormat = game.settings.get("quick-journal-page-clipboard", "outputFormat");
+      const isMarkdown = radioButtonFormat === "markdown";
+      //const returnHTML = radioButtonFormat === "html";
+      const isEmptyLine = game.settings.get("quick-journal-page-clipboard", "isEmptyLineText");
+      let text = await exportJournalText(sheet);
+      if(isMarkdown){
+        text = await convertHtmlToMarkdown(text);
+if(isEmptyLine){
+        text = text.replace(/^\s*\n/gm, "");
+}
+      }
+       
+      await navigator.clipboard.writeText(text); // browser standard for putting text into the clipboard 
+      if(text!==""){ui.notifications.info(game.i18n.localize("QJPC.notifications.copied"));}
+  };
+
+  /*
+  function ensureMutualExclusion() {
+  //const returnMDtext = game.settings.get("quick-journal-page-clipboard", "returnMDtext");
+  //const returnHTMLtext = game.settings.get("quick-journal-page-clipboard", "returnHTMLtext");
+  const radioButtonFormat = game.settings.get("quick-journal-page-clipboard", "outputFormat");
+  const isMarkdown = radioButtonFormat === "markdown";
+  const returnHTML = radioButtonFormat === "html";
+
+  if (isMarkdown && returnHTML) {
+    // If both are true, turn off HTML and keep MD (or vice versa)
+    game.settings.set("quick-journal-page-clipboard", "returnHTMLtext", false);
+    game.settings.set("quick-journal-page-clipboard", "returnMDtext", false);
+    ui.notifications.warn(game.i18n.localize("QJPC.warning.onlyOneAllowed"));
+  }
+}*/